third_party/blink/web_tests/css-parser/deprecated-css-appearance-values-warning/webkit-appearance-warning-cssom-usecounter.html

<!DOCTYPE html>
<link rel=author href="mailto:dizhangg@chromium.org">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<button id="button">foo</button>

<script>
const button = document.getElementById('button');
const useCounterIdNonStandard = 4416;
const useCounterIdInnerSpinButton = 2814;

test((t) => {
  button.style.setProperty('-webkit-appearance', 'auto');
  assert_false(internals.isUseCounted(document, useCounterIdInnerSpinButton));
  assert_false(internals.isUseCounted(document, useCounterIdNonStandard), 'Having an element with -webkit-appearance value "auto" should not activate the use counter (using SetProperty).');

  button.style = '';

  button.style.setProperty('-webkit-appearance', 'inner-spin-button');
  assert_true(internals.isUseCounted(document, useCounterIdInnerSpinButton));
  assert_true(internals.isUseCounted(document, useCounterIdNonStandard), 'Having an element with to be deprecated -webkit-appearance value "inner-spin-button" should activate the use counter (using SetProperty).');
});
</script>
